// src/screens/dashboard/types.ts
type DashboardValuesItem = {
  totalVisitedCount: number;
  totalCollection: number;
  totalAdmission: number;
  totalDischarge: number;
  grandTotalCollection: number;
  totalHospitalCollection: number;
  totalStoreCollection: number;
  totalSamplePending: number;
  totalSampleCollected: number;
  totalResultsPending: number;
  totalResultsDone: number;
};

export type { DashboardValuesItem };